+++ tests/test.rs
@@ -227,3 +227,13 @@ fn sync_once_cell_does_not_leak_partially_constructed_boxes() {
         })
     }
 }
+
+#[test]
+fn unsync_clone() {
+    let s = unsync::OnceCell::new();
+    let c = s.clone();
+    assert!(c.get().is_none());
+    s.set("hello".to_string()).unwrap();
+    let c = s.clone();
+    assert_eq!(c.get().map(String::as_str), Some("hello"));
+}