+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
-## Unreleased
+## 1.0.6
+
+- `unsync::OnceCell<T>` is `Clone` if `T` is `Clone`.
 
 ## 1.0.5
- - No changelog until this point :(
+
+- No changelog until this point :(